+++ src/debugger/util.ts
@@ -31,7 +31,7 @@ function moveTokenCursorToBreakpoint(
     const coor = [...debugResponse.coor]; // Copy the array so we do not modify the one stored in state
 
     for (let i = 0; i < coor.length; i++) {
-        while (!tokenCursor.downList(true)) {
+        while (!tokenCursor.downList()) {
             tokenCursor.next();
         }
         const previousToken = tokenCursor.getPrevToken();